moving some functions around
